// frontend/src/client/types.gen.ts
export type CRMCompanyPublic = {
    name: string;
    website?: (string | null);
    industry?: (string | null);
    employee_count?: (number | null);
    annual_revenue?: (number | null);
    address?: (string | null);
    city?: (string | null);
    state?: (string | null);
    country?: (string | null);
    notes?: (string | null);
    id: string;
    tenant_id: string;
    created_at?: (string | null);
    updated_at?: (string | null);
};